fix: moved migrations
