+++ web/src/SAP/__init__.py
@@ -6,11 +6,49 @@ import openapi_spec_validator
 from flask_cors import CORS
 from jsonschema import RefResolver
 
-from .common.database import get_connection
 from .generated import encoder
 
-SPECIFICATION_DIR = "/app/openapi_specs/SOFI/"
+import sys
+import inspect
+import web.src.SAP.migrations as migrations
+from web.src.SAP.common.database import MIGRATIONS_COL_NAME,DB_NAME, get_connection
 
+
+## --- MIGRATIONS ---
+
+conn = get_connection()
+mydb = conn[DB_NAME]
+migrations_coll = mydb[MIGRATIONS_COL_NAME]
+
+
+# The migrations work by dynamically loading all the functions in the migrations.py file. Then it compares these function names to the ones in the migrations collection
+#  in the database. All migration functions whoose names are not in the migrations collection are executed, and after they are executed they are added to the collection.
+#  This way all migrations are executed exactly once on each database.
+
+already_executed = list(map(lambda x: x["name"], migrations_coll.find()))
+
+migration_funcs = [
+    func for name, func in inspect.getmembers(migrations, inspect.isfunction)
+    if func.__module__ == migrations.__name__  # only functions defined there
+]
+
+for func in migration_funcs:
+    name = func.__name__ 
+
+    if name in already_executed:
+        print("Skipping migration:",name,file=sys.stderr)
+    else:
+        print("Executing migration:",name,file=sys.stderr)
+        try:
+            func()
+        except Exception:
+            print("Migration",name,"failed with error:",sys.exception(),"Stopping all migration until this is fixed!",file=sys.stderr)
+            break
+        migrations_coll.insert({"name": name})
+        
+
+## --- Running the app ---
+SPECIFICATION_DIR = "/app/openapi_specs/SOFI/"
 
 # connexion tries to validate the spec before it bothers to resolve any refs.
 # openapi_spec_validator does not recognize relative $refs, and needs them realized AOT.
